+++ src/Tests/Dfe.Complete.CypressTests/cypress/api/taskHelperConversions.ts
@@ -494,6 +494,50 @@ class TaskHelperConversions extends TaskHelper {
 
         return taskApiConversions.updateTupeConsultationTask(defaultBody);
     }
+
+    updatePrivateFinanceInitiative(taskDataId: string, status: TaskStatus) {
+        const defaultBody = {
+            taskDataId: { value: taskDataId },
+            notApplicable: false,
+            supplementaryFundingAgreementPfiClausesInserted: false,
+            masterFundingAgreementPfiClausesInserted: false,
+            received: false,
+            cleared: false,
+            draftSavedInTrustSharepointFolder: false,
+            signedByAllStakeholders: false,
+            finalVersionSavedInSchoolAndTrustSharepointFolder: false,
+        };
+
+        switch (status) {
+            case "notApplicable":
+                return taskApiConversions.updatePrivateFinanceInitiativeTask({
+                    ...defaultBody,
+                    notApplicable: true,
+                });
+
+            case "inProgress":
+                return taskApiConversions.updatePrivateFinanceInitiativeTask({
+                    ...defaultBody,
+                    supplementaryFundingAgreementPfiClausesInserted: true,
+                });
+
+            case "completed":
+                return taskApiConversions.updatePrivateFinanceInitiativeTask({
+                    taskDataId: { value: taskDataId },
+                    notApplicable: false,
+                    supplementaryFundingAgreementPfiClausesInserted: true,
+                    masterFundingAgreementPfiClausesInserted: true,
+                    received: true,
+                    cleared: true,
+                    draftSavedInTrustSharepointFolder: true,
+                    signedByAllStakeholders: true,
+                    finalVersionSavedInSchoolAndTrustSharepointFolder: true,
+                });
+
+            default:
+                return taskApiConversions.updatePrivateFinanceInitiativeTask(defaultBody);
+        }
+    }
 }
 
 const taskHelperConversions = new TaskHelperConversions();